FIX-#2056: move some utils to modin.utils
