+++ modin/pandas/reshape.py
@@ -15,7 +15,7 @@ import pandas
 
 from .dataframe import DataFrame
 from .series import Series
-from .utils import to_pandas
+from modin.utils import to_pandas
 from modin.error_message import ErrorMessage
